% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{advertising}
\alias{advertising}
\title{Advertising budgets and newspaper sales for 200 markets.}
\format{A data frame with 200 rows and 4 variables:
\describe{
  \item{TV}{TV media budget, in thousands of US dollars}
  \item{Radio}{Radio media budget, in thousands of US dollars}
  \item{Newspaper}{Newspaper media budget, in thousands of US dollars}
  \item{Sales}{unit sales, in thousands}
}}
\source{
\url{http://www-bcf.usc.edu/~gareth/ISL}
}
\usage{
advertising
}
\description{
A dataset containing the advertising budgets and newspaper sales for 200 markets.
}
\keyword{datasets}
