% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tutorial.R
\name{get_tutorial}
\alias{get_tutorial}
\title{Get tutorial}
\usage{
get_tutorial(tutorial)
}
\arguments{
\item{tutorial}{Name of tutorial
\enumerate{
  \item \code{"Hello"}: An introduction
  \item \code{"EDA"}: Exploratory Data Analysis
  \item \code{"Unsupervised"}: Principal Components Analysis & Cluster Analysis
  \item \code{"Linear Regression"}: Linear Regression
  \item \code{"Supervised Classification"}: Logistic Regression & Discriminant Analysis
  \item \code{"Resampling"}: Leave-One-Out Cross-Validation, *k*-Fold Cross Validation, & Bootstrapping
  \item \code{"Model Selection"}: Best Subset & Stepwise Selection for Linear Models
  }}
}
\description{
\code{get_tutorial} will activate the requrested tutorial
}
\examples{
\dontrun{
get_tutorial("Hello")
}

}
